requirements.txt
scrapy
python-crontab
zerorpc
